little modification

+++ 0x00-hello_world/6-size.c
@@ -10,7 +10,7 @@ int main(void)
 {
 	int a;
 	long int b;
-	long long int c;
+	long int c;
 	char d;
 	float f;
